Tried to add a report-making functionality
but llm doesn't seem to use it. will need to change somethings, for
example the llm doesn't see past queries and gets confused. might change
back to structured outputs for simplicity. also llm gets confused for
example by -1's in the ranks, and the database design needs to be
changed so the file name/id is present in the rank. need to explain the
database schema stuff more, maybe copy/pasting the explanations from
darshan-parser.